Move files to utils dir
