+++ examples/trl/ppo.py
@@ -13,9 +13,9 @@ from trl import AutoModelForCausalLMWithValueHead
 from trl.core import LengthSampler
 
 from optimum.habana.accelerate import GaudiAccelerator
-from optimum.habana.environment import disable_kernel
 from optimum.habana.trl import GaudiPPOConfig, GaudiPPOTrainer, adapt_PreTrainedModelWrapper_to_gaudi
 from optimum.habana.utils import set_seed
+from optimum.habana.utils.environment import disable_kernel
 
 
 tqdm.pandas()
