+++ optimum/habana/transformers/models/llama/modeling_llama.py
@@ -28,7 +28,7 @@ from ....distributed.tensorparallel import (
     reduce_from_tensor_model_parallel_region,
 )
 from ....distributed.tp import TPModule
-from ....features import import_usable_component
+from ....utils.features import import_usable_component
 from ...modeling_attn_mask_utils import (
     _gaudi_prepare_4d_causal_attention_mask,
 )
